Convert template to gem
The portability of a template is quite limited. It is difficult to push changes
upstream, and also needs to be added quite early in a project's lifecycle. To
help with that, this commit changes the essentials of the template into a gem.
Projects can add this gem rather than needing the entirety of the template.

For https://tickets.puppetlabs.com/browse/PE-14831
