+++ Rakefile
@@ -1,5 +1,5 @@
 require 'bundler/setup'
 
-require_relative './lib/sinatra-i18n/gettext_setup.rb'
+require_relative './lib/gettext-setup/gettext_setup.rb'
 
 import 'lib/tasks/gettext.rake'
