+++ lib/gettext-setup.rb
@@ -1,0 +1,1 @@
+require 'gettext-setup/gettext_setup'
