+++ locales/config-sample.yaml
@@ -11,6 +11,10 @@ gettext:
   package_name: Sinatra i18n demo
   # The locale that the default messages in the .pot file are in
   default_locale: en
+  # The email used for sending bug reports.
+  bugs_address: docs@puppetlabs.com
+  # The holder of the copyright.
+  copyright_holder: Puppet Labs, LLC.
   # Patterns for +Dir.glob+ used to find all files that might contain
   # translatable content, relative to the project root directory
   source_files:
