--- locales/ja/sinatra-i18n.po
@@ -1,34 +1,0 @@
-# Japanese translations for Sinatra i18n demo package.
-# Copyright (C) 2016 Puppet Labs, LLC.
-# This file is distributed under the same license as the Sinatra i18n demo package.
-# Automatically generated, 2016.
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Sinatra i18n demo \n"
-"Report-Msgid-Bugs-To: docs@puppetlabs.com\n"
-"POT-Creation-Date: 2016-04-05 09:39-0700\n"
-"PO-Revision-Date: 2016-02-26 18:21-0800\n"
-"Last-Translator: Automatically generated\n"
-"Language-Team: none\n"
-"Language: ja\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Plural-Forms: nplurals=1; plural=0;\n"
-
-#: ../app.rb:14
-msgid "Hello, world!"
-msgstr "Hello, world!"
-
-#. Translators need to know some details about the city
-#. So we explain it in a comment
-#: ../app.rb:17
-#, fuzzy
-msgid "There is %{count} bicycle in %{city}"
-msgid_plural "There are %{count} bicycles in %{city}"
-msgstr[0] "%{city}では%{count}の自転車があります。"
-
-#: ../app.rb:18
-msgid "We negotiated a locale of %{locale}"
-msgstr "ロケールとして%{locale}を交渉しました。"
