--- locales/pl/sinatra-i18n.po
@@ -1,36 +1,0 @@
-# Polish translations for Sinatra i18n demo package.
-# Copyright (C) 2016 Puppet Labs, LLC.
-# This file is distributed under the same license as the Sinatra i18n demo package.
-# Automatically generated, 2016.
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Sinatra i18n demo \n"
-"Report-Msgid-Bugs-To: docs@puppetlabs.com\n"
-"POT-Creation-Date: 2016-03-30 11:22-0700\n"
-"PO-Revision-Date: 2016-02-26 18:21-0800\n"
-"Last-Translator: Automatically generated\n"
-"Language-Team: none\n"
-"Language: pl\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Plural-Forms: nplurals=3; plural=(n==1 ? 0 : n%10>=2 && n%10<=4 && (n%100<10 "
-"|| n%100>=20) ? 1 : 2);\n"
-
-#: ../app.rb:14
-msgid "Hello, world!"
-msgstr ""
-
-#. Translators need to know some details about the city here
-#. So we explain it in a comment
-#: ../app.rb:17
-msgid "There is %{count} bicycle in %{city}"
-msgid_plural "There are %{count} bicycles in %{city}"
-msgstr[0] ""
-msgstr[1] ""
-msgstr[2] ""
-
-#: ../app.rb:18
-msgid "We negotiated a locale of %{locale}"
-msgstr ""
