--- locales/sinatra-i18n.pot
@@ -1,36 +1,0 @@
-# SOME DESCRIPTIVE TITLE.
-# Copyright (C) 2016 Puppet Labs, LLC.
-# This file is distributed under the same license as the Sinatra i18n demo package.
-# FIRST AUTHOR <EMAIL@ADDRESS>, 2016.
-#
-#, fuzzy
-msgid ""
-msgstr ""
-"Project-Id-Version: Sinatra i18n demo init-22-gf804f61\n"
-"\n"
-"Report-Msgid-Bugs-To: docs@puppetlabs.com\n"
-"POT-Creation-Date: 2016-04-05 10:42-0700\n"
-"PO-Revision-Date: 2016-04-05 10:42-0700\n"
-"Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
-"Language-Team: LANGUAGE <LL@li.org>\n"
-"Language: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Plural-Forms: nplurals=INTEGER; plural=EXPRESSION;\n"
-
-#: ../app.rb:14
-msgid "Hello, world!"
-msgstr ""
-
-#. Translators need to know some details about the city
-#. So we explain it in a comment
-#: ../app.rb:17
-msgid "There is %{count} bicycle in %{city}"
-msgid_plural "There are %{count} bicycles in %{city}"
-msgstr[0] ""
-msgstr[1] ""
-
-#: ../app.rb:18
-msgid "We negotiated a locale of %{locale}"
-msgstr ""
